refactor: event-types hooks have private and public folders
