+++ packages/platform/atoms/index.ts
@@ -5,8 +5,8 @@ export { BookerPlatformWrapper as Booker } from "./booker";
 export { useIsPlatform } from "./hooks/useIsPlatform";
 export { useAtomsContext } from "./hooks/useAtomsContext";
 export { useConnectedCalendars } from "./hooks/useConnectedCalendars";
-export { useEventTypes } from "./hooks/event-types/useEventTypes";
-export { useEventType as useEvent } from "./hooks/event-types/useEventType";
+export { useEventTypes } from "./hooks/event-types/public/useEventTypes";
+export { useEventType as useEvent } from "./hooks/event-types/public/useEventType";
 export { useEventTypeById } from "./hooks/event-types/private/useEventTypeById";
 export { useCancelBooking } from "./hooks/useCancelBooking";
 export { useGetBooking } from "./hooks/useGetBooking";
